feat: Add FamilySharingScreen widget
This commit adds the `FamilySharingScreen` widget, which is responsible for displaying a screen where users can manage their family sharing settings. The screen includes a button to add family members and a list of existing family members. The data is fetched from Firestore using a stream.

The `FamilySharingScreen` widget takes a `userId` parameter to identify the current user.